respecting shader parameters fixed
There is an option in the celestialbody.gd script that allows of "star light" or system light overrides on the light and ambient light colors in a system. They are commented out at the moment, so it's all defined on the planet library level but it's a nice feature if you want to globally override all the items in the system.