Hyperjump Map improvements
Removed all of the hard coded system positions and made functions that all pull them from the universe.JSON file. Now the creation of systems is simple and in one place.

Additionally made it possible to have systems that have no celestial bodies at all. the system scene simply ignores them when they are not defined.