Fixing the coordinate flipping for Hyperspace jumps
Still need to add the change to the PlayerShip.gd file.

Mission generation is currently impossible because the jumps are suuuuper wrong.

The hyperspace map is greatly improved but still broken in many ways

Will start moving more and more items to the Database, like ship data

Buying ships on Gaia is currently broken too.